Add pipelines (see desc), history and a WhichWrapper sans which lookup
The four pipeline classes are briefly described in the Pipeline javadoc. PipeArg.Pipeable is to be implemented in types that need to get piped around (more of these to come). PipeCMD is to be implemented in the classes that actually do things, and pipelines are set up to pass PipeArg<>s to and from PipeCMDs.